# packages/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Oris API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/encounters/{encounter_id}/audio/chunks/{sequence}": {
      "put": {
        "tags": [
          "audio"
        ],
        "summary": "Put Chunk",
        "description": "Idempotent : renvoyer le même segment est sans effet.",
        "operationId": "put_chunk_encounters__encounter_id__audio_chunks__sequence__put",
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Encounter Id"
            }
          },
          {
            "name": "sequence",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sequence"
            }
          },
          {
            "name": "content-type",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content-Type"
            }
          },
          {
            "name": "x-chunk-timestamp-ms",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "X-Chunk-Timestamp-Ms"
            }
          },
          {
            "name": "x-chunk-checksum",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 64,
              "maxLength": 64,
              "title": "X-Chunk-Checksum"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "audio/pcm;rate=16000;channels=1;encoding=s16le": {
              "schema": {
                "type": "string",
                "contentMediaType": "application/octet-stream",
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkReceiptOut"
                }
              }
            }
          },
          "200": {
            "description": "Segment déjà reçu à l'identique"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/encounters/{encounter_id}/audio/gaps": {
      "post": {
        "tags": [
          "audio"
        ],
        "summary": "Report Gap",
        "operationId": "report_gap_encounters__encounter_id__audio_gaps_post",
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Encounter Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioGapReport"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioSessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/encounters/{encounter_id}/audio": {
      "get": {
        "tags": [
          "audio"
        ],
        "summary": "Get Audio",
        "operationId": "get_audio_encounters__encounter_id__audio_get",
        "parameters": [
          {
            "name": "encounter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Encounter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioSessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioGapOut": {
        "properties": {
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "duration_ms"
        ],
        "title": "AudioGapOut"
      },
      "AudioGapReport": {
        "properties": {
          "reason": {
            "type": "string",
            "enum": [
              "microphone_lost",
              "page_reloaded",
              "capture_error"
            ],
            "title": "Reason"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "AudioGapReport"
      },
      "AudioSessionOut": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "audio_format": {
            "type": "string",
            "title": "Audio Format"
          },
          "received_count": {
            "type": "integer",
            "title": "Received Count"
          },
          "last_sequence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sequence"
          },
          "next_sequence": {
            "type": "integer",
            "title": "Next Sequence"
          },
          "next_timestamp_ms": {
            "type": "integer",
            "title": "Next Timestamp Ms"
          },
          "missing_sequences": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Missing Sequences"
          },
          "received_duration_ms": {
            "type": "integer",
            "title": "Received Duration Ms"
          },
          "gaps": {
            "items": {
              "$ref": "#/components/schemas/AudioGapOut"
            },
            "type": "array",
            "title": "Gaps"
          },
          "reported_gap_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reported Gap Reasons"
          },
          "purge_status": {
            "type": "string",
            "title": "Purge Status"
          },
          "finalized_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finalized At"
          },
          "purged_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purged At"
          },
          "last_received_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Received At"
          }
        },
        "type": "object",
        "required": [
          "status",
          "audio_format",
          "received_count",
          "last_sequence",
          "next_sequence",
          "next_timestamp_ms",
          "missing_sequences",
          "received_duration_ms",
          "gaps",
          "reported_gap_reasons",
          "purge_status",
          "finalized_at",
          "purged_at",
          "last_received_at"
        ],
        "title": "AudioSessionOut"
      },
      "ChunkReceiptOut": {
        "properties": {
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "status": {
            "type": "string",
            "enum": [
              "stored",
              "duplicate"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "sequence",
          "status"
        ],
        "title": "ChunkReceiptOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}